+++ test/preface_qcheck/gen.ml
@@ -23,6 +23,15 @@ let nonempty_list g s = nonempty_list_size nat g s
 
 let small_nonempty_list g s = nonempty_list_size small_nat g s
 
+let seq_size size gen st =
+  let init = Preface_stdlib.Seq.pure (gen st) in
+  fold (fun acc _ -> Seq.cons (gen st) acc) init (size st)
+;;
+
+let seq g s = seq_size nat g s
+
+let small_seq g s = seq_size small_nat g s
+
 let continuation f state = Preface_stdlib.Continuation.pure (f state)
 
 let exn state =