+++ test/preface_stdlib_test/preface_stdlib_test.ml
@@ -5,6 +5,7 @@ let () =
     @ Continuation_test.cases
     @ List_test.cases
     @ Nonempty_list_test.cases
+    @ Seq_test.cases
     @ Option_test.cases
     @ State_test.cases
     @ Stream_test.cases